Move form to separate component
